MNT-21837 - License expires 23h earlier - correction (#544) (#547)
* period.getMillis() (joda lib) returns an int instead of a long, so on larger periods (like 30days) we get an error that the value cannot fit in an int
* Changed the validation to only calculate the remaining milliseconds in the edge case of the remaining days being zero (as both 23h before and 23h after correspond to 0 remaining days)
* Added a unit test that validates the usage 60 days up from the licence expiring to validate the problem stops occurring
* Added a unit test that validates the usage 30 days after the licence expiring to cover all use cases

(cherry picked from commit 6e1d25a8f082171ccf65e393554ccf8a8c2dfc8e)

+++ repository/src/main/java/org/alfresco/util/DateUtil.java
@@ -80,12 +80,12 @@ public class DateUtil
      * @param endMs   end date in milliseconds
      * @return number milliseconds between
      */
-    public static long calculateMs(long startMs, long endMs)
+    public static int calculateMs(long startMs, long endMs)
     {
         DateTime startDateTime = new DateTime(startMs);
         DateTime endDateTime = new DateTime(endMs);
 
-        long milliseconds;
+        int milliseconds;
         if (endDateTime.isBefore(startDateTime))
         {
             Interval interval = new Interval(endDateTime, startDateTime);
